requirements.txt
# Core dependencies
mss>=9.0.1
Pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.14.0

# Testing
pytest>=8.0.0
